refactor: moved log to sails helper
