+++ tests/unit/fixtures/gameSocketEvent.js
@@ -46,22 +46,18 @@ const resolveThreeData = {
         id: 1,
         playedBy: 0,
         moveType: 'resolveThree',
-        playedCard: Card.THREE_OF_SPADES,
-        targetCard: Card.TEN_OF_HEARTS,
-        discardedCard: null,
+        playedCard: '3S',
+        targetCard: 'TH',
+        discardedCards: [],
         turn: 1,
-        phase: 3,
-        p0: {
-          hand: [Card.TEN_OF_HEARTS],
-          points: [],
-          faceCards: [],
-        },
-        p1: {
-          hand: [Card.TEN_OF_DIAMONDS],
-          points: [Card.ACE_OF_HEARTS],
-          faceCards: [],
-        },
-        deck: [Card.SIX_OF_SPADES, Card.TWO_OF_HEARTS, Card.ACE_OF_CLUBS],
+        phase: 1,
+        p0Hand: ['TH'],
+        p0Points: [],
+        p0FaceCards: [],
+        p1Hand: ['TD'],
+        p1Points: ['AH'],
+        p1FaceCards: [],
+        deck: ['6S', '2H', 'AC'],
         scrap: [],
         oneOff: null,
         oneOffTarget: null,
@@ -81,9 +77,9 @@ const resolveThreeData = {
     moveType: 'resolveThree',
     playedCard: Card.THREE_OF_SPADES,
     targetCard: Card.TEN_OF_HEARTS,
-    discardedCard: null,
+    discardedCards: [],
     turn: 1,
-    phase: 3,
+    phase: 1,
     p0: {
       hand: [Card.TEN_OF_HEARTS],
       points: [],
@@ -95,6 +91,31 @@ const resolveThreeData = {
       faceCards: [],
     },
     deck: [Card.SIX_OF_SPADES, Card.TWO_OF_HEARTS, Card.ACE_OF_CLUBS],
+    scrap: [],
+    oneOff: null,
+    oneOffTarget: null,
+    twos: [],
+    resolving: null,
+    gameId: 9,
+    createdAt: '2024-07-14T12:59:47.709Z',
+  },
+
+  gameStateRow: {
+    id: 1,
+    playedBy: 0,
+    moveType: 'resolveThree',
+    playedCard: '3S',
+    targetCard: 'TH',
+    discardedCards: [],
+    turn: 1,
+    phase: 1,
+    p0Hand: ['TH'],
+    p0Points: [],
+    p0FaceCards: [],
+    p1Hand: ['TD'],
+    p1Points: ['AH'],
+    p1FaceCards: [],
+    deck: ['6S', '2H', 'AC'],
     scrap: [],
     oneOff: null,
     oneOffTarget: null,
